+++ crates/api/CHANGELOG.md
@@ -12,6 +12,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `crate::Annotation` now wraps `String`s inside `Term`s.
 - New type: `crate::term::Builder`. Intended for internal use only, but had to make public to
   use it in the `term!()` macro.
+- Added `parse_unit!()` macro as a shortcut to `Unit::from_str(foo).unwrap()`.
 
 ### Changed